feat(oem-fe): Location data added to form and message entity
Longitude and latitude as well as a radius can now be entered for a
message to indicate the validity for a certain region. The form is also
validated to not allow coordinate values out of boundaries and
incomplete coordinates. The message format has been updated and is
correctly persisted in and retrieved from the db.

BREAKING CHANGE: Location data added

Closes #56