feat(android): Incoming messages now can be persisted
If a new message is received by the amqp-client, it can be persisted by
selecting that within the message-view. When the location favorites are
added, there will be a differentiation, if the message will be
automatically persisted, too.

Additionally, the components for reading
all messages from the database and deleting elements was added too.

BREAKING CHANGE: Incoming messages now can be persisted