feat(Android): Location favourites now can be created and managed
Location favourites can now be created in a separate view. They get
persisted locally in a Room DB. The location can be received by locating
the device or manual entering of an address, which then is
validated.
Autocomplete Feature was not implemented due to api
costs.
Geofencing part is still missing.

BREAKING CHANGE: Location favourites now can be created and managed